learning-platform/backend/src/requirements.txt
flask
pymongo
python-dotenv
flask-caching
flask-login
flask-jwt-extended
openai
numpy
scipy
werkzeug 
